--- plugins/baser-core/tests/TestCase/View/Helper/_NotYetMigrated/BcFormTableHelperTest.php
@@ -1,39 +1,0 @@
-<?php
-// TODO : コード確認要
-return;
-/**
- * baserCMS :  Based Website Development Project <https://basercms.net>
- * Copyright (c) baserCMS Users Community <https://basercms.net/community/>
- *
- * @copyright       Copyright (c) baserCMS Users Community
- * @link            https://basercms.net baserCMS Project
- * @package         Baser.Test.Case.View.Helper
- * @since           baserCMS v 3.0.6
- * @license         https://basercms.net/license/index.html
- */
-
-App::uses('BcAppView', 'View');
-App::uses('BcFormTableHelper', 'View/Helper');
-
-class BcFormTableHelperTest extends BcTestCase
-{
-    public function setUp()
-    {
-        parent::setUp();
-    }
-
-    public function tearDown()
-    {
-        parent::tearDown();
-    }
-
-    public function testDispatchBefore()
-    {
-        $this->markTestIncomplete('このテストは、まだ実装されていません。');
-    }
-
-    public function testDispatchAfter()
-    {
-        $this->markTestIncomplete('このテストは、まだ実装されていません。');
-    }
-}
